requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
cssselect==1.3.0
dataclasses-json==0.6.7
ddgs==9.0.0
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
duckduckgo_search==8.1.1
executing==2.2.0
fastapi==0.115.12
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
google_search_results==2.4.2
googlenewsdecoder==0.1.7
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.61
langchain-openai==0.3.18
langchain-perplexity==0.1.1
langchain-tavily==0.2.7
langchain-text-splitters==0.3.8
langgraph==0.4.5
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.0
langgraph-sdk==0.1.70
langsmith==0.3.42
lxml==5.4.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.40.0
nest-asyncio==1.6.0
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.6
openai==1.82.0
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
primp==0.15.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pymongo==4.13.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rpds-py==0.25.1
scikit-learn==1.6.1
scipy==1.15.3
selectolax==0.3.29
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
streamlit==1.45.1
tabulate==0.9.0
tavily-python==0.7.2
tenacity==9.1.2
textblob==0.19.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tinysegmenter==0.3
tldextract==5.3.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
watchdog==6.0.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.0
zipp==3.23.0
zstandard==0.23.0

# Document generation libraries
python-docx==1.1.2
reportlab==4.0.8
markdown==3.5.2
beautifulsoup4==4.13.4
